Start work on subdirectory support
Issue #12

+++ inc/security.php
@@ -13,12 +13,12 @@ include APPPATH . 'config.php';
 
 // Check if MySQL has been set and the install folder exists so we can redirect safely.
 if ($mysql_host == '' && file_exists(__DIR__ . "/../pages/install/")) {
-  header("Location: /pages/install/install.php", true, 302);
+  header("Location: " . $_SERVER['REQUEST_URI'] . "pages/install/install.php", true, 302);
   die();
 }
 
 // If MySQL is not set and install folder is missing, return an error.
-if ($mysql_host == '') {
+if (!isset($mysql_host) && !file_exists(__DIR__ . "/../pages/install/")) {
   echo "<p style='color: #f00; padding: 20px;'>Something is not right! The install folder is not in the normal location and the configurations file has not been correctly setup!</p>";
   die();
 }
